fix two categorical var color issues: 1.) seemingly arbitrarily incorrect color mappings for integers, fixed with 0.5 offset for all except min and max (should revisit), and 2.) incorrect color order for color map (should revisit messy fix)

+++ drilldown/drill_log.py
@@ -45,7 +45,7 @@ def clean_missing_intervals(depths, values):
 
 def convert_fractional_rgb_to_rgba_for_plotly(rgb, opacity=1):
     return "rgba({},{},{}, {})".format(
-        int(rgb[0] * 255), int(rgb[1] * 255), int(rgb[2] * 255), opacity
+        rgb[0] * 255, rgb[1] * 255, rgb[2] * 255, opacity
     )
